Paso 34: Arreglados test rotos,cambiado de nombre de la pagina principal, refactorizacion del evento y del listener, mejora del comando para borrar anuncios viejos , creados nuevos tests y arreglos varios

+++ app/Http/Controllers/AdvertisementController.php
@@ -50,6 +50,7 @@ class AdvertisementController extends Controller
 
     public function edit(Advertisement $advertisement)
     {
+        $this->authorize('update', $advertisement);
         return view('advertisements.edit', compact('advertisement'));
     }
 
@@ -74,6 +75,7 @@ class AdvertisementController extends Controller
 
     public function destroy(Advertisement $advertisement)
     {
+        $this->authorize('delete', $advertisement);
         $advertisement->delete();
 
         return to_route('welcome')
